% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpgui-dependencies.R
\name{add_rpgui_deps}
\alias{add_rpgui_deps}
\title{rpgui dependencies utils}
\usage{
add_rpgui_deps(tag)
}
\arguments{
\item{tag}{Element to attach the dependencies.}
}
\description{
This function attaches rpgui. dependencies to the given tag
}
