% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyRPG-dependencies.R
\name{add_shinyRPG_deps}
\alias{add_shinyRPG_deps}
\title{shinyRPG dependencies utils}
\usage{
add_shinyRPG_deps(tag)
}
\arguments{
\item{tag}{Element to attach the dependencies.}
}
\description{
This function attaches shinyRPG dependencies to the given tag
}
