% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{rpgButton}
\alias{rpgButton}
\title{Create a button input}
\usage{
rpgButton(inputId, label, golden = FALSE)
}
\arguments{
\item{inputId}{\link{rpgSlider} unique id.}

\item{label}{Text writtent on top of the input.}

\item{golden}{Style parameter. If TRUE, the slider is bordered by fancy elements.
FALSE by default.}
}
\description{
\link{rpgButton} creates a button input element.
}
