% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{rpgCheckbox}
\alias{rpgCheckbox}
\alias{updateRpgCheckbox}
\title{Create a checkbox input}
\usage{
rpgCheckbox(inputId, label, value = FALSE, golden = FALSE)

updateRpgCheckbox(
  session = getDefaultReactiveDomain(),
  inputId,
  label = NULL,
  value = NULL
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{value}{Initial value (\code{TRUE} or \code{FALSE}).}

\item{session}{The \code{session} object passed to function given to
\code{shinyServer}. Default is \code{getDefaultReactiveDomain()}.}
}
\description{
\link{rpgCheckbox} is a nice checkbox input with RPG design.

\link{updateRpgCheckbox} allows to update a \link{rpgCheckbox} on the server.
}
\note{
For now, only the selected value may be updated.
}
