% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{rpgProgress}
\alias{rpgProgress}
\alias{updateRpgProgress}
\title{Create a progress bar}
\usage{
rpgProgress(id, value, color = "purple")

updateRpgProgress(
  id,
  value = NULL,
  color = NULL,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{id}{\link{rpgSlider} unique id.}

\item{value}{New value.}

\item{color}{New color.}

\item{session}{Shiny session object.}
}
\description{
\link{rpgProgress} creates a simple progress bar.

\link{updateRpgProgress} changes the progress bar configuration from the server.
}
