% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{rpgSelect}
\alias{rpgSelect}
\alias{updateRpgSelect}
\title{Create a select input}
\usage{
rpgSelect(
  inputId,
  label,
  choices,
  selected = NULL,
  multiple = FALSE,
  size = NULL
)

updateRpgSelect(
  session = getDefaultReactiveDomain(),
  inputId,
  label = NULL,
  choices = NULL,
  selected = NULL
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{choices}{List of values to select from. If elements of the list are
named, then that name --- rather than the value --- is displayed to the
user. It's also possible to group related inputs by providing a named list
whose elements are (either named or unnamed) lists, vectors, or factors. In
this case, the outermost names will be used as the group labels (leveraging
the \verb{<optgroup>} HTML tag) for the elements in the respective sublist. See
the example section for a small demo of this feature.}

\item{selected}{The initially selected value (or multiple values if \code{multiple = TRUE}). If not specified then defaults to the first value for
single-select lists and no values for multiple select lists.}

\item{multiple}{Is selection of multiple items allowed?}

\item{size}{Number of items to show in the selection box; a larger number
will result in a taller box. Not compatible with \code{selectize=TRUE}.
Normally, when \code{multiple=FALSE}, a select input will be a drop-down list,
but when \code{size} is set, it will be a box instead.}

\item{session}{The \code{session} object passed to function given to
\code{shinyServer}. Default is \code{getDefaultReactiveDomain()}.}
}
\description{
\link{rpgSelect} creates a dropdown select input. Whenever size is not NULL,
it is displayed as a box.

\link{updateRpgSelect} allows to update a \link{rpgSelect} on the server.
}
\note{
For now, only the selected value may be updated.
}
