% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{rpgSlider}
\alias{rpgSlider}
\alias{updateRpgSlider}
\title{Create a slider input}
\usage{
rpgSlider(inputId, label, min, max, value, golden = FALSE)

updateRpgSlider(session = shiny::getDefaultReactiveDomain(), inputId, value)
}
\arguments{
\item{inputId}{\link{rpgSlider} unique id.}

\item{label}{Text writtent on top of the input.}

\item{min}{The minimum value (inclusive) that can be selected.}

\item{max}{The maximum value (inclusive) that can be selected.}

\item{value}{\link{rpgSlider} value.}

\item{golden}{Style parameter. If TRUE, the slider is bordered by fancy elements.
FALSE by default.}

\item{session}{Shiny session object.}
}
\description{
\link{rpgSlider} creates a slider input element.

\link{updateRpgSlider} changes the slider value from the server.
}
