% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{rpgCenter}
\alias{rpgCenter}
\title{Utility to center elements}
\usage{
rpgCenter(tag)
}
\arguments{
\item{tag}{Tag on which to apply the class.}
}
\value{
The input tag with modified class.
}
\description{
Utility to center elements
}
