% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{rpgContainer}
\alias{rpgContainer}
\title{Beautiful container}
\usage{
rpgContainer(..., style = NULL)
}
\arguments{
\item{...}{Ui elements.}

\item{style}{Style. NULL by default. Valid choices are framed, framed-golden,
framed-golden-2 and framed-grey. See \link{validRpgStyles}.}
}
\description{
Container to include RPG elements inside.
}
