% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{rpgDisable}
\alias{rpgDisable}
\title{Utility to disable elements}
\usage{
rpgDisable(tag)
}
\arguments{
\item{tag}{Tag on which to apply the attribute.}
}
\value{
The input tag with modified attribute.
}
\description{
Utility to disable elements
}
