% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{shinyRPGDemo}
\alias{shinyRPGDemo}
\title{Launch the shinyRPG Gallery}
\usage{
shinyRPGDemo()
}
\description{
A gallery of all components available in shinyRPG.
}
\examples{

if (interactive()) {

 shinyRPGDemo()

}
}
